# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is PBL VI main gateway"
  version: "1.0.0"
  title: "Petfeeder Gateway"
  contact:
    email: "isacartur@gmail.com"
host: "localhost"
basePath: "/v1"
tags:
  - name: "user"
    description: "A person using mobile applicatin"
  - name: "device"
    description: "Raspberry Pi device inside of petfeeder"
  - name: "pet"
    description: "Animal which petfeeder will serve"
schemes:
  - "https"
  - "http"
paths:
  /user/register:
    post:
      tags:
        - "user"
      summary: "Register a new user"
      description: ""
      operationId: "registerUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Credential for a new user"
          required: true
          schema:
            $ref: "#/definitions/UserCredentials"
      responses:
        200:
          description: "Successfully created a user"
        401:
          description: "Invalid email or email"
          schema:
            $ref: "#/definitions/Error"
  /user/login:
    post:
      tags:
        - "user"
      summary: "Login a new user"
      description: ""
      operationId: "loginUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Credential to try to authenticate user"
          required: true
          schema:
            $ref: "#/definitions/UserCredentials"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/Token"
        401:
          description: "Unauthenticated"
          schema:
            $ref: "#/definitions/Error"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      uuid:
        type: "string"
      mail:
        type: "string"
        description: "Must be a valid email"
      password:
        type: "string"
        description: "Must be 8 characters long. Include at least 1 special character, at least 1 digit and at least one upper case character"
  UserCredentials:
    type: "object"
    properties:
      mail:
        type: "string"
        description: "Must be a valid email"
      password:
        type: "string"
        description: "Must be 8 characters long. Include at least 1 special character, at least 1 digit and at least one upper case character"
  Token:
    type: "object"
    properties:
      token:
        type: "string"
        description: "HMAC JWT token"
  Error:
    type: "object"
    properties:
      error:
        type: "string"
        description: "Error message"